Reference counting for TagLib::File related objects
As TagLib::File does destroy all underlying TagLib::Tag and TagLib::Frame
objects, we have to make sure that we don't destroy our file object
before releasing every tag and frame object.
To do this we separate the file object handling  and definitions
from the other objects to have a clean refcounting in file_objects,
and references to the file object in the tag and frame objects.

We use taglib_ref_class after every initialisation of a file related object
to setup the necessary references, which are dropped in the Taglib_free method.
The taglib_ref_class function is specified in taglibrary.h as it uses
typdefs from the C++ headers.

Signed-off-by: David Soria Parra <dsp@php.net>