Return false if there is no tag instead of throwing an exception
Scripts might call getID3v1Tag to determine if there is a ID3v1 tag
at all. If we throw an exception we will force the developer to
use catch statements for his script flow.

Therefore we return just false instead of throwing an exception, so that
the developer can (as expected) change his workflow based on the return
value instead of an exception.

Signed-off-by: David Soria Parra <dsp@php.net>